changed wrapper parameters, corrected CMakeLists
- changed the wrapper parameter types of y and y-hat from double to int since they represent nominal/ordinal data
- added `prims/add_sub_dev_scalar.cu` to CMakeLists.txt

+++ cpp/src/metrics/metrics.cu
@@ -39,7 +39,7 @@ namespace ML {
             return MLCommon::Metrics::computeRandIndex(y, y_hat, (uint64_t)n, handle.getDeviceAllocator(), handle.getStream());
         }
 
-        double adjustedRandIndex(const cumlHandle& handle, double *y, double *y_hat, int n, double lower_class_range, double upper_class_range){
+        double adjustedRandIndex(const cumlHandle& handle,  int *y,  int *y_hat, int n, int lower_class_range,  int upper_class_range){
             return MLCommon::Metrics::computeAdjustedRandIndex(y, y_hat, n, lower_class_range, upper_class_range, handle.getDeviceAllocator(), handle.getStream());
         }
